Improve onboarding, AI summaries, and article detail view
- Refactor onboarding into extension-based architecture with separate files per step
- Add inline feed discovery as final onboarding step with duplicate detection
- Update localization terminology from "add/subscribe" to "follow" across all languages
- Add transition animations to AI summary cards (fade in/out)
- Fix debug force-visible summaries to respect dismiss state
- Split today's summary generation into time-windowed batches to avoid context limits
- Fix article summary cache not loading for WebView-extracted articles
- Show summarization errors in alert instead of silently failing
- Add blur replace transitions for article translation and summary toggling
- Filter hidden elements and .sosumi class in WebView content extraction
- Clean up database schema to v1 (remove migration code)
- Resolve all needs_review localization strings
